Revert "move fungiable asset yransfer method to fungible asset client class"
This reverts commit 740f8311912d2ab8393075ae98c3d02f0e11be0d.
